Merge FCL-Team:"main" to "24.2.0"

+++ FCL/src/main/java/com/tungsten/fcl/ui/manage/ManagePage.java
@@ -7,15 +7,19 @@ import android.widget.ScrollView;
 
 import com.tungsten.fcl.R;
 import com.tungsten.fcl.setting.Profile;
+import com.tungsten.fcl.ui.ProgressDialog;
 import com.tungsten.fcl.ui.UIManager;
 import com.tungsten.fcl.ui.version.Versions;
 import com.tungsten.fcl.util.RequestCodes;
+import com.tungsten.fclauncher.utils.FCLPath;
 import com.tungsten.fclcore.fakefx.beans.property.BooleanProperty;
 import com.tungsten.fclcore.fakefx.beans.property.SimpleBooleanProperty;
+import com.tungsten.fclcore.task.Schedulers;
 import com.tungsten.fclcore.task.Task;
 import com.tungsten.fclcore.util.io.FileUtils;
 import com.tungsten.fcllibrary.browser.FileBrowser;
 import com.tungsten.fcllibrary.browser.options.LibMode;
+import com.tungsten.fcllibrary.component.dialog.FCLAlertDialog;
 import com.tungsten.fcllibrary.component.theme.ThemeEngine;
 import com.tungsten.fcllibrary.component.ui.FCLCommonPage;
 import com.tungsten.fcllibrary.component.view.FCLImageButton;
@@ -31,6 +35,7 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
     private ScrollView left;
     private ScrollView right;
 
+    private FCLImageButton browseFCLLog;
     private FCLImageButton browseGame;
     private FCLImageButton browseMod;
     private FCLImageButton browseConfig;
@@ -70,6 +75,7 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
         ThemeEngine.getInstance().registerEvent(left, () -> left.setBackgroundTintList(new ColorStateList(new int[][] { { } }, new int[] { ThemeEngine.getInstance().getTheme().getLtColor() })));
         ThemeEngine.getInstance().registerEvent(right, () -> right.setBackgroundTintList(new ColorStateList(new int[][] { { } }, new int[] { ThemeEngine.getInstance().getTheme().getLtColor() })));
 
+        browseFCLLog = findViewById(R.id.browse_fcl_logs);
         browseGame = findViewById(R.id.browse_game_dir);
         browseMod = findViewById(R.id.browse_mods);
         browseConfig = findViewById(R.id.browse_config);
@@ -85,6 +91,7 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
         redownload = findViewById(R.id.update_assets);
         deleteLibs = findViewById(R.id.delete_libs);
         deleteLogs = findViewById(R.id.delete_logs);
+        browseFCLLog.setOnClickListener(this);
         browseGame.setOnClickListener(this);
         browseMod.setOnClickListener(this);
         browseConfig.setOnClickListener(this);
@@ -105,10 +112,11 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
         updateLayout.visibilityProperty().bind(currentVersionUpgradable);
     }
 
-    private void onBrowse(String sub) {
+    private void onBrowse(String dir) {
         FileBrowser.Builder builder = new FileBrowser.Builder(getContext());
         builder.setLibMode(LibMode.FILE_BROWSER);
-        builder.setInitDir(new File(getProfile().getRepository().getRunDirectory(getVersion()), sub).getAbsolutePath());
+        dir = dir.startsWith("/") ? dir : new File(getProfile().getRepository().getRunDirectory(getVersion()), dir).getAbsolutePath();
+        builder.setInitDir(dir);
         builder.create().browse(getActivity(), RequestCodes.BROWSE_DIR_CODE, null);
     }
 
@@ -117,11 +125,35 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
     }
 
     private void clearLibraries() {
-        FileUtils.deleteDirectoryQuietly(new File(getProfile().getRepository().getBaseDirectory(), "libraries"));
+        FCLAlertDialog.Builder builder = new FCLAlertDialog.Builder(getContext());
+        builder.setAlertLevel(FCLAlertDialog.AlertLevel.ALERT);
+        builder.setMessage(String.format(getContext().getString(R.string.version_manage_remove_confirm), "libraries"));
+        builder.setPositiveButton(() -> {
+            ProgressDialog progress = new ProgressDialog(getContext());
+            Task.runAsync(() -> {
+                FileUtils.deleteDirectoryQuietly(new File(getProfile().getRepository().getBaseDirectory(), "libraries"));
+            }).whenComplete(Schedulers.androidUIThread(), (e) -> {
+                progress.dismiss();
+            }).start();
+        });
+        builder.setNegativeButton(null);
+        builder.create().show();
     }
 
     private void clearJunkFiles() {
-        Versions.cleanVersion(getProfile(), getVersion());
+        FCLAlertDialog.Builder builder = new FCLAlertDialog.Builder(getContext());
+        builder.setAlertLevel(FCLAlertDialog.AlertLevel.ALERT);
+        builder.setMessage(String.format(getContext().getString(R.string.version_manage_remove_confirm), "logs"));
+        builder.setPositiveButton(() -> {
+            ProgressDialog progress = new ProgressDialog(getContext());
+            Task.runAsync(() -> {
+                Versions.cleanVersion(getProfile(), getVersion());
+            }).whenComplete(Schedulers.androidUIThread(), (e) -> {
+                progress.dismiss();
+            }).start();
+        });
+        builder.setNegativeButton(null);
+        builder.create().show();
     }
 
     private void updateGame() {
@@ -151,6 +183,9 @@ public class ManagePage extends FCLCommonPage implements ManageUI.VersionLoadabl
 
     @Override
     public void onClick(View view) {
+        if (view == browseFCLLog) {
+            onBrowse(FCLPath.LOG_DIR);
+        }
         if (view == browseGame) {
             onBrowse("");
         }
